requirements.txt
gensim
pymongo
scrapy
goose-extractor